requirements.txt
pandas
numpy
scikit-learn
joblib
dice-ml

# Multi-agent debate system
autogen-agentchat>=0.4.9
autogen-ext[openai]>=0.4.9
python-dotenv>=1.0.1